// src/app/model/galaxias.ts
export interface Galaxias {

    id: Number
    nome: String;
    urlImg: String;
    magnitude: Number;
    declinacao: String;
    distancia: String;
    desvioVermelho: String;
    descricao: String
    rotacao: Number;
    dimensoes: Number;
    raio: Number;
    IdadeEstimada: Number;
    massa: Number;
    estrelas: Number;
    constelacao: String;
    coordenadas: String;

}